+++ demos/chinook/src/main/java/is/codion/demos/chinook/domain/api/Chinook.java
@@ -63,7 +63,7 @@ public interface Chinook {
 	DomainType DOMAIN = domainType(Chinook.class);
 
 	interface Artist {
-		EntityType TYPE = DOMAIN.entityType("chinook.artist", Artist.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.artist", Artist.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> NAME = TYPE.stringColumn("name");
@@ -72,7 +72,7 @@ public interface Chinook {
 	}
 
 	interface Album {
-		EntityType TYPE = DOMAIN.entityType("chinook.album", Album.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.album", Album.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> TITLE = TYPE.stringColumn("title");
@@ -89,7 +89,7 @@ public interface Chinook {
 	}
 
 	interface Employee {
-		EntityType TYPE = DOMAIN.entityType("chinook.employee", Employee.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.employee", Employee.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> LASTNAME = TYPE.stringColumn("lastname");
@@ -114,7 +114,7 @@ public interface Chinook {
 	}
 
 	interface Customer {
-		EntityType TYPE = DOMAIN.entityType("chinook.customer", Customer.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.customer", Customer.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> FIRSTNAME = TYPE.stringColumn("firstname");
@@ -139,7 +139,7 @@ public interface Chinook {
 	}
 
 	interface Preferences {
-		EntityType TYPE = DOMAIN.entityType("chinook.preferences", Preferences.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.preferences", Preferences.class);
 
 		Column<Long> CUSTOMER_ID = TYPE.longColumn("customer_id");
 		Column<Long> PREFERRED_GENRE_ID = TYPE.longColumn("preferred_genre_id");
@@ -150,21 +150,21 @@ public interface Chinook {
 	}
 
 	interface Genre {
-		EntityType TYPE = DOMAIN.entityType("chinook.genre", Genre.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.genre", Genre.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> NAME = TYPE.stringColumn("name");
 	}
 
 	interface MediaType {
-		EntityType TYPE = DOMAIN.entityType("chinook.mediatype", MediaType.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.mediatype", MediaType.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> NAME = TYPE.stringColumn("name");
 	}
 
 	interface Track {
-		EntityType TYPE = DOMAIN.entityType("chinook.track", Track.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.track", Track.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> NAME = TYPE.stringColumn("name");
@@ -203,7 +203,7 @@ public interface Chinook {
 	}
 
 	interface Invoice {
-		EntityType TYPE = DOMAIN.entityType("chinook.invoice", Invoice.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.invoice", Invoice.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<Long> CUSTOMER_ID = TYPE.longColumn("customer_id");
@@ -229,7 +229,7 @@ public interface Chinook {
 	}
 
 	interface InvoiceLine {
-		EntityType TYPE = DOMAIN.entityType("chinook.invoiceline", InvoiceLine.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.invoiceline", InvoiceLine.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<Long> INVOICE_ID = TYPE.longColumn("invoice_id");
@@ -246,7 +246,7 @@ public interface Chinook {
 	}
 
 	interface Playlist {
-		EntityType TYPE = DOMAIN.entityType("chinook.playlist", Playlist.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.playlist", Playlist.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<String> NAME = TYPE.stringColumn("name");
@@ -259,7 +259,7 @@ public interface Chinook {
 	}
 
 	interface PlaylistTrack {
-		EntityType TYPE = DOMAIN.entityType("chinook.playlisttrack", PlaylistTrack.class.getName());
+		EntityType TYPE = DOMAIN.entityType("chinook.playlisttrack", PlaylistTrack.class);
 
 		Column<Long> ID = TYPE.longColumn("id");
 		Column<Long> PLAYLIST_ID = TYPE.longColumn("playlist_id");