+++ framework/domain/src/main/java/is/codion/framework/domain/DomainType.java
@@ -45,6 +45,15 @@ public sealed interface DomainType permits DefaultDomainType {
 	 * Instantiates a {@link EntityType} associated with this domain type.
 	 * If this entity type has been defined previously that instance is returned.
 	 * @param name the entity type name
+	 * @param resourceBundleClass the class to use as base name of the resource bundle to use for captions
+	 * @return a {@link EntityType} with the given name
+	 */
+	EntityType entityType(String name, Class<?> resourceBundleClass);
+
+	/**
+	 * Instantiates a {@link EntityType} associated with this domain type.
+	 * If this entity type has been defined previously that instance is returned.
+	 * @param name the entity type name
 	 * @param resourceBundleName the name of a resource bundle to use for captions
 	 * @return a {@link EntityType} with the given name
 	 */